+++ src/promptflow-tools/promptflow/tools/common.py
@@ -16,7 +16,6 @@ from promptflow.contracts.types import PromptTemplate
 from promptflow.exceptions import SystemErrorException, UserErrorException
 from promptflow.tools.exception import (
     ToolValidationError,
-    ChatAPIAssistantRoleInvalidFormat,
     ChatAPIFunctionRoleInvalidFormat,
     ChatAPIToolRoleInvalidFormat,
     ChatAPIInvalidFunctions,
@@ -222,7 +221,7 @@ def validate_tools(tools):
 def try_parse_name_and_content(role_prompt):
     # customer can add ## in front of name/content for markdown highlight.
     # and we still support name/content without ## prefix for backward compatibility.
-    pattern = r"\n*#{0,2}\s*name:\n+\s*(\S+)\s*\n*#{0,2}\s*content:\n?(.*)"
+    pattern = r"\n*#{0,2}\s*name\s*:\s*\n+\s*(\S+)\s*\n*#{0,2}\s*content\s*:\s*\n?(.*)"
     match = re.search(pattern, role_prompt, re.DOTALL)
     if match:
         return match.group(1), match.group(2)
@@ -232,7 +231,7 @@ def try_parse_name_and_content(role_prompt):
 def try_parse_tool_call_id_and_content(role_prompt):
     # customer can add ## in front of tool_call_id/content for markdown highlight.
     # and we still support tool_call_id/content without ## prefix for backward compatibility.
-    pattern = r"\n*#{0,2}\s*tool_call_id:\n+\s*(\S+)\s*\n*#{0,2}\s*content:\n?(.*)"
+    pattern = r"\n*#{0,2}\s*tool_call_id\s*:\s*\n+\s*(\S+)\s*\n*#{0,2}\s*content\s*:\s*\n?(.*)"
     match = re.search(pattern, role_prompt, re.DOTALL)
     if match:
         return match.group(1), match.group(2)
@@ -242,36 +241,19 @@ def try_parse_tool_call_id_and_content(role_prompt):
 def try_parse_tool_calls(role_prompt):
     # customer can add ## in front of tool_calls for markdown highlight.
     # and we still support tool_calls without ## prefix for backward compatibility.
-    pattern = r"\n*#{0,2}\s*tool_calls:\n*\s*(\[.*?\])"
+    pattern = r"\n*#{0,2}\s*tool_calls\s*:\s*\n+\s*(\[.*?\])"
     match = re.search(pattern, role_prompt, re.DOTALL)
     if match:
-        return match.group(1)
+        try:
+            parsed_array = eval(match.group(1))
+            return parsed_array
+        except Exception:
+            None
     return None
 
 
-def is_tools_chunk(last_message):
+def is_tool_chunk(last_message):
     return last_message and "role" in last_message and last_message["role"] == "tool" and "content" not in last_message
-
-
-def is_assistant_tool_calls_chunk(last_message, chunk):
-    return last_message and "role" in last_message and last_message["role"] == "assistant" and "tool_calls" in chunk
-
-
-def parse_tool_calls_for_assistant(last_message, chunk):
-    parsed_result = try_parse_tool_calls(chunk)
-    error_msg = "Failed to parse assistant role prompt with tool_calls. Please make sure the prompt follows the format:"
-    " 'tool_calls:\\n[{ id: tool_call_id, type: tool_type, function: {name: function_name, arguments: function_args }]'"
-    "See more details in https://platform.openai.com/docs/api-reference/chat/create#chat-create-messages"
-
-    if parsed_result is None:
-        raise ChatAPIAssistantRoleInvalidFormat(message=error_msg)
-    else:
-        parsed_array = None
-        try:
-            parsed_array = eval(parsed_result)
-            last_message["tool_calls"] = parsed_array
-        except Exception:
-            raise ChatAPIAssistantRoleInvalidFormat(message=error_msg)
 
 
 def parse_tools(last_message, chunk, hash2images, image_detail):
@@ -311,13 +293,15 @@ def parse_chat(
 
     for chunk in chunks:
         last_message = chat_list[-1] if len(chat_list) > 0 else None
-        if is_tools_chunk(last_message):
+        if is_tool_chunk(last_message):
             parse_tools(last_message, chunk, hash2images, image_detail)
             continue
 
-        if is_assistant_tool_calls_chunk(last_message, chunk):
-            parse_tool_calls_for_assistant(last_message, chunk)
-            continue
+        if last_message and "role" in last_message and last_message["role"] == "assistant":
+            parsed_result = try_parse_tool_calls(chunk)
+            if parsed_result is not None:
+                last_message["tool_calls"] = parsed_result
+                continue
 
         if (
             last_message