+++ src/promptflow-core/CHANGELOG.md
@@ -1,6 +1,11 @@
 # promptflow-core package
 
-## v1.10.0 (Upcoming)
+## v1.10.0 (2024.04.26)
+
+### Features Added
+- Add prompty feature to simplify the development of prompt templates for customers, reach [here](https://microsoft.github.io/promptflow/how-to-guides/develop-a-prompty/index.html) for more details.
+
+### Others
 - Add fastapi serving engine support.
 
 ## v1.9.0 (2024.04.17)